pyproject.toml
[build-system]
requires = ["hatchling>=1.20.0"]
build-backend = "hatchling.build"

[project]
name = "gardener"
version = "0.1.0"
description = "Automated static analysis and dependecy graph centrality analysis for recommendation of OSS dependency funding splits to be distributed via the Drips Network"
readme = "README.md"
requires-python = ">=3.9,<3.12"
authors = [
  { name = "Warren Winter" }
]
license = { file = "LICENSE" }

dependencies = [
  "GitPython==3.1.44",
  "pathspec==0.12.1",
  "grep-ast==0.8.1",
  "networkx==3.2.1",
  "requests==2.32.3",
  "pydantic>=2,<3",
  "numpy>=1.23,<2",
  "scipy==1.11.4",
  "toml",
  "tomli; python_version < '3.11'",
  "tqdm",
  "tenacity",
]

[project.optional-dependencies]
viz = [
  "ipysigma>=0.21",
]
service = [
  "tree-sitter==0.23.2",
  "tree-sitter-language-pack==0.6.1",
  "fastapi==0.104.1",
  "uvicorn[standard]==0.24.0",
  "celery==5.3.4",
  "redis==5.0.1",
  "sqlalchemy==2.0.23",
  "psycopg2-binary==2.9.9",
  "alembic==1.13.0",
  "python-dotenv==1.0.0",
  "pydantic-settings==2.1.0",
  "python-multipart==0.0.6",
  "httpx==0.25.2",
  "slowapi==0.1.9",
]
test = [
  "pytest>=7.4",
  "pytest-mock>=3.11",
  "psutil>=5.9",
  "pytest-timeout>=2.3",
  "pyyaml>=6.0",
]
dev = [
  "tox>=4",
  "tox-uv>=1.6.0",
  "pylint>=3.0",
  "mypy>=1.7.0",
  "black>=24.3.0",
  "isort>=5.12.0",
]

[project.scripts]
gardener = "gardener.main_cli:main"

[tool.hatch.build]
include = [
  "gardener/**",
  "README.md",
]
exclude = [
  "gardener/external_helpers/**/node_modules/**",
  "gardener/external_helpers/**/package-lock.json",
  "gardener/external_helpers/**/.DS_Store",
]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 120
force_single_line = false
combine_as_imports = true

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
ignore_missing_imports = true
no_site_packages = true
pretty = true
allow_redefinition = true
follow_imports = "silent"
# The codebase is not typed; keep mypy lenient to provide value without breaking CI
disable_error_code = "truthy-function,assignment,valid-type,misc,import"
